+++ src/notebook/index.ts
@@ -7,9 +7,12 @@ import { CodeTour } from "../store";
 import { getStepFileUri, getWorkspaceUri } from "../utils";
 
 class CodeTourNotebookProvider implements vscode.NotebookSerializer {
-  originalContent: Uint8Array = new TextEncoder().encode('');
+  originalContent: Uint8Array = new TextEncoder().encode("");
 
-  async deserializeNotebook(content: Uint8Array, token: any): Promise<vscode.NotebookData> {
+  async deserializeNotebook(
+    content: Uint8Array,
+    token: any
+  ): Promise<vscode.NotebookData> {
     this.originalContent = content;
     let contents = new TextDecoder().decode(content);
 
@@ -21,8 +24,8 @@ class CodeTourNotebookProvider implements vscode.NotebookSerializer {
       const uri = await getStepFileUri(item, workspaceRoot, tour.ref);
       const document = await vscode.workspace.openTextDocument(uri);
 
-      const startLine = (item.line! > 10) ? item.line! - 10 : 0;
-      const endLine = (item.line! > 1) ? item.line! - 1 : 0;
+      const startLine = item.line! > 10 ? item.line! - 10 : 0;
+      const endLine = item.line! > 1 ? item.line! - 1 : 0;
       const contents = document.getText(
         new vscode.Range(
           new vscode.Position(startLine, 0),
@@ -40,24 +43,39 @@ class CodeTourNotebookProvider implements vscode.NotebookSerializer {
     let cells: vscode.NotebookCellData[] = [];
 
     // Title cell
-    cells.push(new vscode.NotebookCellData(1,
-      `## ![Icon](${SMALL_ICON_URL})&nbsp;&nbsp; CodeTour (${tour.title}) - ${steps.length} steps\n\n${tour.description === undefined ? '' : tour.description}`,
-      'markdown'))
+    cells.push(
+      new vscode.NotebookCellData(
+        1,
+        `## ![Icon](${SMALL_ICON_URL})&nbsp;&nbsp; CodeTour (${tour.title}) - ${
+          steps.length
+        } steps\n\n${tour.description === undefined ? "" : tour.description}`,
+        "markdown"
+      )
+    );
 
     steps.forEach((step, index) => {
-      cells.push(new vscode.NotebookCellData(2,
-        step.contents,
-        step.language,
-        [new vscode.NotebookCellOutput([
-          new vscode.NotebookCellOutputItem('text/markdown', `_Step #${index + 1} of ${steps.length}:_ ${step.description} ([View File](${step.uri}))`)
-        ])]
-      ))
-    })
+      const cell = new vscode.NotebookCellData(2, step.contents, step.language);
+      cell.outputs = [
+        new vscode.NotebookCellOutput([
+          new vscode.NotebookCellOutputItem(
+            new TextEncoder().encode(
+              `_Step #${index + 1} of ${steps.length}:_ ${
+                step.description
+              } ([View File](${step.uri}))`
+            ),
+            "text/markdown"
+          )
+        ])
+      ];
+    });
 
-    return new vscode.NotebookData(cells, new vscode.NotebookDocumentMetadata(true))
+    return new vscode.NotebookData(cells);
   }
 
-  async serializeNotebook(data: vscode.NotebookData, token: any): Promise<Uint8Array> {
+  async serializeNotebook(
+    data: vscode.NotebookData,
+    token: any
+  ): Promise<Uint8Array> {
     return this.originalContent;
   }
 }